requirements.txt
# Gmail API
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# AI/LLM
anthropic>=0.40.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Environment Management
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication & Security
flask-login==0.6.3
flask-bcrypt==1.0.1
cryptography==41.0.7

# Caching
redis==5.0.1

# Utilities
requests==2.31.0
